triage-backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "triage-backend"
version = "1.0.0"
description = "TRIAGE — Multi-Agent Hospital Crisis Simulation Environment for Meta PyTorch OpenEnv Hackathon"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "TRIAGE Team"},
]
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",

    # Database
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",

    # Redis & Celery
    "redis>=5.1.0",
    "celery[redis]>=5.4.0",

    # AI / ML
    "transformers>=4.46.0",
    "huggingface_hub>=0.26.0",
    "torch>=2.5.0",
    "datasets>=3.0.0",
    "trl>=0.12.0",
    "peft>=0.13.0",
    "accelerate>=1.0.0",
    "bitsandbytes>=0.44.0",

    # Utilities
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "pyyaml>=6.0.2",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "websockets>=13.0",
    "python-dotenv>=1.0.1",
    "rank-bm25>=0.2.2",
    "openenv-core==0.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "mypy>=1.11.0",
    "ruff>=0.7.0",
]
unsloth = [
    "unsloth>=2024.12",
]
postgres = [
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.9",
]

[project.scripts]
triage-server = "triage.api.main:main"
triage-episode = "scripts.run_episode:main"
triage-train = "scripts.run_training:main"
triage-report = "scripts.generate_training_report:main"

[tool.setuptools.packages.find]
include = ["triage*", "config*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
